refactor(console): console/PluginManager
