+++ console/src/pages/Settings/PluginManager/index.module.less
@@ -8,7 +8,6 @@
 .content {
   flex: 1;
   overflow: auto;
-  padding: 24px;
 }
 
 .table {
